Created A* search algorithm
This is a A* Path finding algorithm which is visual. You will be able to see the algorithm work its way to the end goal. You will need to install a module called pygame for this to work. However, this is really easy. You will be able to move the walls around so you can make it different each time. There is also ways you can change this and potentially make it even more effective. I think it is good it can be expanded on because it means other people can learn and make it better. This algorithm is more challenging but it is also very interesting.  I learnt this algorithm from a youtuber called tech with tim. Here is the link to the video for further explanation. 
https://www.youtube.com/watch?v=JtiK0DOeI4A